+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.1 under development
 
-- no changes in this release.
+- Enh #4: Add the `no_empty_statement` rule to `@Yiisoft/Core` (@mspirkov)
 
 ## 1.0.0 January 23, 2026