+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1 under development
 
 - Enh #4: Add the `no_empty_statement` rule to `@Yiisoft/Core` (@mspirkov)
+- Enh #10: Add `RemoveOverrideAttributeRector` rule (@vjik)
 
 ## 1.0.0 January 23, 2026