+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Yii Code Style Change Log
 
-## 1.0.1 under development
+## 1.1.0 July 27, 2026
 
-- New #4: Add the `no_empty_statement` rule to `@Yiisoft/Core` (@mspirkov)
 - New #11: Add `RemoveOverrideAttributeRector` Rector rule (@vjik)
 - New #11: Add `SetList::YII_CORE` Rector set (@vjik)
+- New #4: Add the `no_empty_statement` rule to `@Yiisoft/Core` (@mspirkov)
 
 ## 1.0.0 January 23, 2026